moved activitys added arabic translation file
