+++ app/src/main/AndroidManifest.xml
@@ -5,13 +5,13 @@
 	<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 	<uses-permission android:name="android.permission.CAMERA" />
 	<application android:usesCleartextTraffic="true" android:allowBackup="true" android:icon="@mipmap/ic_launcher" android:label="@string/app_name" android:roundIcon="@mipmap/ic_launcher_round" android:supportsRtl="true" android:theme="@style/AppTheme">
-		<activity android:name="com.hr.techlabapp.Statistics" android:screenOrientation="portrait" />
-		<activity android:name="com.hr.techlabapp.AddProduct" android:screenOrientation="portrait" />
+		<activity android:name="com.hr.techlabapp.Activitys.Statistics" android:screenOrientation="portrait" />
+		<activity android:name="com.hr.techlabapp.Activitys.AddProduct" android:screenOrientation="portrait" />
 		<!-- TODO: Add labels -->
-		<activity android:name="com.hr.techlabapp.CheckLendRequest" android:screenOrientation="portrait" />
-		<activity android:name="com.hr.techlabapp.ProductList" android:screenOrientation="portrait" />
-		<activity android:name="com.hr.techlabapp.ProductInfo" android:screenOrientation="portrait" />
-		<activity android:name="com.hr.techlabapp.MainActivity" android:label="@string/app_name" android:screenOrientation="portrait" android:theme="@style/AppTheme.NoActionBar">
+		<activity android:name="com.hr.techlabapp.Activitys.CheckLendRequest" android:screenOrientation="portrait" />
+		<activity android:name="com.hr.techlabapp.Activitys.ProductList" android:screenOrientation="portrait" />
+		<activity android:name="com.hr.techlabapp.Activitys.ProductInfo" android:screenOrientation="portrait" />
+		<activity android:name="com.hr.techlabapp.Activitys.MainActivity" android:label="@string/app_name" android:screenOrientation="portrait" android:theme="@style/AppTheme.NoActionBar">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
 				<action android:name="android.intent.action.VIEW" />
