+++ app/src/main/java/com/hr/techlabapp/Activitys/AddProduct.java
@@ -1,7 +1,9 @@
-package com.hr.techlabapp;
+package com.hr.techlabapp.Activitys;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.hr.techlabapp.R;
 
 public class AddProduct extends AppCompatActivity {
