+++ app/src/main/java/com/hr/techlabapp/Activitys/DeleteItemDialog.java
@@ -1,4 +1,4 @@
-package com.hr.techlabapp;
+package com.hr.techlabapp.Activitys;
 
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
