+++ app/src/main/java/com/hr/techlabapp/Activitys/cGrid.java
@@ -1,15 +1,12 @@
-package com.hr.techlabapp;
+package com.hr.techlabapp.Activitys;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.res.Resources;
-import android.graphics.Point;
 import android.os.Build;
 import android.support.annotation.Nullable;
 import android.support.annotation.RequiresApi;
 import android.util.AttributeSet;
 import android.util.DisplayMetrics;
-import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.View;
