+++ app/src/main/res/layout/activity_add_product.xml
@@ -4,7 +4,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".AddProduct">
+	tools:context=".Activitys.AddProduct">
 	<!---->
 	<TextView
 		android:id="@+id/product_name_text"
