+++ app/src/main/res/layout/activity_check_lend_request.xml
@@ -4,7 +4,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".CheckLendRequest"
+	tools:context=".Activitys.CheckLendRequest"
 	android:padding="@dimen/usual_margin">
 	<Space
 		android:id="@+id/user_history_spacer"
