+++ app/src/main/res/layout/activity_product_info.xml
@@ -4,7 +4,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".ProductInfo">
+	tools:context=".Activitys.ProductInfo">
 	<ImageView
 		android:id="@+id/image"
 		app:layout_constraintTop_toTopOf="parent"
