+++ app/src/main/res/layout/activity_statistics.xml
@@ -4,7 +4,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".Statistics">
+	tools:context=".Activitys.Statistics">
 	<android.support.constraint.ConstraintLayout
 		android:id="@+id/product_statistics"
 		app:layout_constraintTop_toTopOf="parent"
