+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.hr.techlabapp.MainActivity">
+    tools:context="com.hr.techlabapp.Activitys.MainActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
